Fix default Unicode alternate skip count
According to the Word2007 RTF specs page 18 for control word \ucN, the
default value is to skip one byte after a Unicode character. I was
reading weird ‚?‘ characters after German Umlauts, until I understood
what the problem is. This RTF does not contain the \ucN control word
anywhere, but relies on the default.