Initial document model and document building parser
There is a package „document“ which defines a bunch of basic interfaces
which represent aspects of an RTF document. DocumentPart refers to an
object that can contain paragraphs of styled text. There are currently
three areas of the document supported: The Header, Footer and the
document itself. The document is divided into Sections, which is a
concept supported by RTF and many text editing applications. Global
formatting properties can change at section breaks. Furthermore, the
Document contains a ColorTable, FontTable, StyleSheet and
DocumentSettings. The idea is that code that works with the document
model gets an instance of a Document and can get instances of all the
other interfaces from the document instance.
There is an implementation of all interfaces in the package
document.impl. The classes sometimes offer additional functionality,
but in the last iterations of changes, I have eliminated the need for
that by moving functionality into the interfaces.
In the package parser.builder, there is a new RtfListener
implementation called DocumentBuilder. The DocumentBuilder is created
with an instance of Document. Then it is passed to the
StandardParser.parse() method.
The idea behind DocumentBuilder is that it maintains a stack of
RtfContext instances and delegates all events to the currently active
context. RtfContext has an interface almost exactly like RtfListener,
except there is an additional processGroupStart() method version which
has the same parameters as processCommand(). The reason is that some
RTF groups start with a specific command and denote an RTF
„destination“. DocumentBuilder has therefore the notion of a „delayed
group start“. In processGroupStart() it will only mark that it is
currently at a group start, but does nothing further yet. In
processCommand() it will try to form a command based group start when
it encounters a destination command. Otherwise it will just process the
delayed group start before processing any other events.
Various RtfContext implementations exist which can already process
events to build many parts of the document model.