Work on StyleSheet support
Style is now split up into a base interface Style. It defines all
properties and Styles can have a name and return the overridden
properties.
The first derived interface is CharacterStyle and it defines setters
and getters for just the character style related properties.
The second derived interface is ParagraphStyle which also inherits
CharacterStyle. It has an additional method for creating a derived
CharacterStyle that has a ParagraphStyle as its parent.
StyleSheet no longer directly holds styles. Rather it defines getters
for instances of CharacterStyleTable and ParagraphStyleTable. The 2007
spec also knows table and section style tables which are not yet
implemented.
Work in DocumentBuilder towards implementing parsing the styles in the
style sheet section. Currently it only creates the styles, sets their
names and puts them into the correct table with their correct ID.
Also new is ignoring unknown groups. These are detected by a group
directly followed by an Command.optionalcommand, i.e. a command not in
the Command table. The complete group is ignored by the Document
Builder, which fixes some unwanted text ending up in the document.
LibreOffice apparently uses a new version of the spec which has some
section descriptions that we don’t understand which is no longer a
problem.